Remove --use-deprecated legacy-resolver flag
The pip team is planning to depreciate the
'--use-deprecated legacy-resolver'install option.
While a firm date has yet to be set, We need to be
ready for the change.

Test Plan*:

PASS: Verified that tox runs without issues
PASS: Generate Vault package.
PASS: Upload and apply Vault package and check if pod
      is running.

Note(*): Tested on kubernetes 1.21.8

Closes-bug:1966076

Signed-off-by: Karla Felix <karla.karolinenogueirafelix@windriver.com>
Change-Id: I582378ff198430c182fd9197976e9041f2910680