add option to delay server unseal
This change delays unsealing recovering vault servers for 15 seconds.

vault-manager automatically unseals vault servers in a perpetual loop
after initial configuration.  This final loop is the HA recovery
procedure.

It is observed that vault-manager will unseal a recovering vault server
when the active server has not started to send heartbeats to the new
pod.  The result is that the recovering server will timeout waiting for
heartbeats and start an election.  Although the active and standby
server will reject the election, there being a leader already, the
recovering vault will increment 'term' and restart election until
heartbeats are received, or until it wins election.

Although the raft algorithm is resilient to this, the procedure is not
appropriate.  It is better to unseal the vault server after the active
vault sends heartbeats to the new pod.

It is observed that the heartbeat interval reduces promptly from less
than 1 second per heartbeat to ~10-12 seconds for a failed vault server.
So it is reasonable for vault-manager to wait 12 seconds before
unsealing the recovering vault.  This also assumes the vault-manager and
active vault server would receive updated pod information at about the
same time and the latest heartbeat was issued immediately prior to the
update.

The options are configurable in helm-overrides.  The defaults for
example:
  manager:
    statusCheckRate: 5
    unsealWaitIntervals: 3

statusCheckRate is the rate at which vault-manager will check pod
status, in seconds.  unsealWaitIntervals is the number of intervals to
wait before unsealing the server.

Default is 5 s/interval * 3 intervals == 15 seconds

When unsealWaitIntervals is set to 0 there is no delay in unsealing the
recovering vault servers.  This is equivalent to the existing behaviour
before this change when statusCheckRate is also set to 5, which is the
value hard-coded before this change.

Test Plan:
PASS - vault ha 3 replicas
PASS - vault 1 replica
PASS - kubectl exec kill vault process
PASS - kubectl delete vault pod
PASS - short network downtime
PASS - long network downtime
PASS - rates including 1, 5
PASS - wait intervals including 0, 1, 3, 15
PASS - not reproduced with default values (many attempts)

Story: 2010393
Task: 47701

Change-Id: I763f6becee3e1a17e838a4f8ca59b2b0d33ba639
Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>