adjust server readiness probe
It is observed that vault pods consistently show readiness probe warning
when applying the application or when a pod is recovering.  The probe
runs "vault status" which returns failure when the vault is sealed.  The
probe failure is not impactful, but since there is a certain delay
before unseal is completed, adjust initialDelaySeconds to 25 to account
for the time required to unseal vault pods.  This commit should usually
omit readiness probe warning for a single recovering vault pod.

During testing it is observed that:
  Setting initialDelaySeconds to 15: a recovering pod shows readiness
probe warning.
  Setting initialDelaySeconds to 18, a recovering pod omits readiness
probe warning.
  On application-apply, the first pod to be unsealed _may_ show readiness
probe warning when initialDelaySeconds is 25.  Other pods will be
unsealed serially and will show readiness probe warning.

Test Plan:
  PASS  Standard controller storage 2+2
  PASS  HA tests, log inspection
  PASS  Inspection of kubectl describe of pods with various values for
        initialDelaySeconds

Story: 2010393
Task: 48237

Depends-On: https://review.opendev.org/c/starlingx/vault-armada-app/+/884553

Change-Id: I9ea6cca2b591c40bfe70737c0fb390b18b69f796
Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>