exit_on_trap: adjust default and debug behaviours
It is observed that the exit_on_trap is not working under normal
operation - vault-manager takes 30 seconds to exit when the application
is removed.  The default behaviour is to exit when the trap file exists.
Not exiting when the trap file exists only occurs when the trap file has
content.  This latter behaviour is a debugging option.

Add the conditional for empty trap file, and always exit if the trap
file is empty.

For the debugging feature it is helpful for the procedure to remember
the trap number set in the trap file.  Use DEBUGGING_TRAP global
variable to remember the debugging trap requested, and exit whenever
that exit_on_call trap is run.

Also:
Refactor the parameter variable as 'trap' for readability. Adjust all
the logs for exit_on_trap to permit search for "exit_on_trap".  And log
at INFO level when exiting vault-manager.

Test Plan:
PASS  unit test exit_on_trap
PASS  default behavior; vault-manager responds promptly to termination
PASS  debug feature exits on matching trap (remembers debug trap)
PASS  bashate on the rendered init.sh

Story: 2010930
Task: 48843

Change-Id: Id67a89e063daa18ba7627553ac2a19ca673ff00b
Signed-off-by: Michel Thebeau <Michel.Thebeau@windriver.com>