Fix KUBE_LATEST_VERSION to installed version
After updating the kubectl versions, we noticed that there is a mismatch
between version in the KUBE_LATEST_VERSION variable, and the actual
installed version. This change fixes the mismatch, and makes
KUBE_LATEST_VERSION point to the correct version.

Test Plan:
PASS Manual build of the image
PASS verify in docker history for correct version

Story: 2010930
Task: 49526

Change-Id: I5055fd204527c49cc47478d62d01e1afa18d3556
Signed-off-by: Tae Park <tae.park@windriver.com>