mount-helper: run as root
When performing upgrade, conversion of the PVC storage to k8s secrets
requires the helper pod to run as root in order to delete the files in
the volume.  The old vault-manager pod was running as root.

Commit c849a3bb did not account for this and did not test
application-update.

Test Plan:
PASS  platform, application upgrade
PASS  unit test, manual execution of the procedure with pvc-attach.yaml

Story: 2011073
Task: 50522

Change-Id: I9ab340544408e9478e6fce68d78337d2fded9a09
Signed-off-by: Michel Thebeau <Michel.Thebeau@windriver.com>